requirements.txt
streamlit>=1.39.0
pandas>=2.2.2
numpy>=1.26.4
plotly>=6.1.1
scikit-learn>=1.5.0
openpyxl>=3.1.5
reportlab>=4.2.2
google-generativeai>=0.4.0
python-dotenv>=1.0.1
darkdetect>=0.8.0
kaleido==0.2.1